+++ CHANGELOG.md
@@ -5,6 +5,15 @@
 * stocks
 * delete toggle :problem/avail true/false, htmx.
 * htmx from problems page
+* list answers
+* add comments
+
+
+# 0.2.5-SNAPSHOT (2025-09-14)
+
+- upload answers
+* tasks/div-answers do not wrap.
+
 
 # 0.2.4 (2025-09-14)